Move image to a separate file
